Updated table definitions.
Added event_type and event_date fields to attendance table.
The event_date field is calculated from the timestamp. The
event_type will be 'meeting' for regular robotics meetings.
There is a multi-field unique constraint on date, event_type,
and student_id, so a student can't attend two events of the same
event_type in a sngle day.

Incremented version to 0.2.0.