Removed open connections.
Several methods in database.py were not explicitly closing their
connections. This wasn't causing a problem when the app was running,
but it was causing tests to fail when they were all run together,
because connections remaining open were causing permission errors that
were preventing preparations for the next test.