Changed attendance table name to checkins.
All checks and tests pass.
Incremented version to 0.4.0.

A checkin occurs when a student scans their QR code. A checkin
consists of a student_id, timestamp, and event type.

Renamed attendance table because their isn't a simple singular
form of the word "attendance," which was making it hard to name
things.